[PRODUCTO-472] Se modifica el identificador asociado a los documentos del grid para que sea único, ya que al asignar identificadores iguales entre pestañas se marcaban los mismos checks en todas.
Former-commit-id: 72912ea2821d9eca9881771a5c1f192c518e5135

+++ recovery-precontencioso-plugin/src/main/java/es/pfsgroup/plugin/precontencioso/documento/assembler/DocumentoAssembler.java
@@ -30,7 +30,7 @@ public class DocumentoAssembler {
 	 */
 	public static SolicitudDocumentoPCODto docAndSolEntityToSolicitudDto(DocumentoPCO documento,
 			SolicitudDocumentoPCO solicitud, String ugIdDto, String descripcionUG, boolean esDocumento, 
-			boolean tieneSolicitud, DDSiNo siNo, DDSiNoNoAplica siNoAplica, int idIdentificativo) {
+			boolean tieneSolicitud, DDSiNo siNo, DDSiNoNoAplica siNoAplica, Long idIdentificativo) {
 		SolicitudDocumentoPCODto solicitudDto = new SolicitudDocumentoPCODto();
 				
 		solicitudDto.setIdIdentificativo(idIdentificativo + "");
